requirements.txt
pillow==9.5.0
tkinterdnd2
